% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{mb_index_load}
\alias{mb_index_load}
\title{Load a native minibwa index}
\usage{
mb_index_load(prefix, meth = FALSE)
}
\arguments{
\item{prefix}{Index prefix.}

\item{meth}{Load the methylation BWT index.}
}
\value{
An external pointer handle to a native minibwa index.
}
\description{
Load a native minibwa index
}
