% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{minibwa_available}
\alias{minibwa_available}
\title{Test whether the minibwa CLI is available}
\usage{
minibwa_available(path = Sys.getenv("RMINIBWA_MINIBWA", unset = "minibwa"))
}
\arguments{
\item{path}{Executable name or path.}
}
\value{
\code{TRUE} when an executable can be resolved, otherwise \code{FALSE}.
}
\description{
Test whether the minibwa CLI is available
}
