% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{minibwa_version}
\alias{minibwa_version}
\title{Report the minibwa CLI version}
\usage{
minibwa_version(path = minibwa_path())
}
\arguments{
\item{path}{Optional executable name or path. When \code{NULL}, use Rminibwa's
packaged minibwa executable.}
}
\value{
A scalar character version string.
}
\description{
Report the minibwa CLI version
}
