% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{minibwa_version}
\alias{minibwa_version}
\title{Report the minibwa CLI version}
\usage{
minibwa_version(path = minibwa_path())
}
\arguments{
\item{path}{Executable name or path.}
}
\value{
A scalar character version string.
}
\description{
Report the minibwa CLI version
}
