% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Rminibwa-package}
\alias{Rminibwa}
\alias{Rminibwa-package}
\title{Rminibwa: Native R bindings for minibwa}
\description{
Rminibwa wraps the upstream \code{minibwa} short-read aligner with native C
bindings, raw-vector query input, external-pointer alignment batches, ALTREP
column views, downstream C-callable accessors, and runtime-dispatched KSW
SIMD backends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/Rminibwa}
  \item \url{https://sounkou-bioinfo.github.io/Rminibwa/}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/Rminibwa/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Dana-Farber Cancer Institute (Copyright holder of upstream minibwa) [copyright holder]
  \item Ilya Grebnov (libsais copyright holder) [copyright holder]
  \item N. Jesper Larsson (QSufSort copyright holder) [copyright holder]
  \item Wong Chi Kwong (BWT-SW/bwtgen copyright holder) [copyright holder]
}

}
\keyword{internal}
