% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Rminibwa-package}
\alias{Rminibwa}
\alias{Rminibwa-package}
\title{Rminibwa: R Interface and Runtime Dispatch Experiments for minibwa}
\description{
Rminibwa is an R-facing package for exploring command-line and native-library
integration with the upstream \code{minibwa} short-read aligner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/Rminibwa}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/Rminibwa/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Dana-Farber Cancer Institute (Copyright holder of upstream minibwa, when vendored or linked) [copyright holder]
}

}
\keyword{internal}
